test_many_algorithms add finite case for binding bounds, precision lookup dict for failing tests, move nag_dfols test to nag_optimizers, move test_pygmo_optimizers
