+++ tests/optimagic/optimizers/test_pygmo_optimizers.py
